Move node balancer into cluster package
Initial work for #3377.
